Style

+++ src/irvine/oeis/a027/A027834.java
@@ -20,7 +20,7 @@ public class A027834 implements Sequence {
           return Z.ONE;
         }
         final Z zn = Z.valueOf(n);
-        Z sum = zn.pow(r * n);
+        Z sum = zn.pow(r * (long) n);
         for (int t = 1; t < n; ++t) {
           sum = sum.subtract(Binomial.binomial(n, t).multiply(zn.pow(r * (long) (n - t))).multiply(get(r, t)));
         }
